requirements.txt
anthropic>=0.30.0
python-dotenv>=1.0.0
textual>=0.50.0
claude-agent-sdk>=0.1.0
